DO NOT MERGE: Patch to enable products on v4.9
Current sync point in v4.9 kernel does not have the product files and
modifies the primary mdp interface. Fixing locally for now, proper fix
will come in as part of a merge up to the tip of QC's kernel.

This change will be reverted once we merge up to the latest version of
the kernel.

Bug: 111221277
Test: Verified that all devices boot into screen properly.
Change-Id: Ia60c7086a766c9ba33876a0d70b3491d5bb549d9