ARM: msm: dts: enable gpio configuration of ti amp
To power on the external ti amplifier, gpio68 is used.
On OpenQ 624 Platform, gpio68 is getting configured as part of
cdc_pdm_comp_lines instead of cdc_quat_mi2s_gpios.
cdc_pdm_comp_lines are not applicable for OpenQ 624 Platform
due to which gpio68 always getting pulled to low resulting in
ti amplifier not getting powered on.

Bug: 111411689
Test: Verified audio playback.
Change-Id: I5d76a0a1418035728c6a1ca3a2f7dc8aecdcb446
Signed-off-by: Vijay Navnath Kamble <vkambl@codeaurora.org>