qcacld-2.0: fix possible OOB read in SIOCSIWRATE WEXT ioctl handler
propagation from qcacld-3.0 to qcacld-2.0.
The values used to dictate the end of the for loop is great than the
size of array supp_rates, this will cause an OOB read when loop
through supp_rates. So need modify the size of array supp_rates.
There's also a functional issue in that the second call to
sme_cfg_get_str() overwrites the lower values of the first call,
thus not ever allowing the lower channel rates of A to ever be
valid. So need update the read buffer address for the second
sme_cfg_get_str().

Change-Id: Ie73462c2aed21e5321c5d016302ef2a0828aba3b
CRs-Fixed: 2257156
Test: None
Bug: 111906763