Modify qpnp-power-on to support gpio sysfs
The SPMI RESIN interrupt line is responsible for triggering the fastboot
command and doubles as a VOLUME_DOWN pin. However, in certain use-cases,
we want to suppress these keyevents and instead directly read the line
state (high/low) from user-space. In order to support this, we add
functionality here to expose this interrupt line as a virtual GPIO.

The feature is guarded by the config QPNP_POWER_ON_GPIO and only takes
effect if the device-tree sets the `resin-gpiobase` entry. A virtual
gpiochip is created with the provided base, with a single GPIO that
reflects the interrupt state. This GPIO should be treated as a read-only
pin, and will reject any attempts to configure itself as an output pin.

In this change, we rely on passing a simulated interrupt line back to
the GPIO subsystem to support edge-based triggers. The logic for this
line has been backported from mainline Linux's `irq_sim` files.

This change is a cherry-pick of ag/3573118, modified for this kernel.

Bug: 72956949
Test: Test app to get GPIO callbacks on both edge triggers works. State
is reflected properly, even when pressing the button rapidly.
Change-Id: Ied13c5ddc825317d7b68e0871db4db5c77e32852