qcacld-2.0: Use request manager for linkspeed
We are transitioning to the new request manager framework.
Change wlan_hdd_get_linkspeed_for_peermac() to this framework.
Note that this framework provides the infrastructure to pass data
from the response thread to the request thread and hence
eliminates the need to maintain tSirLinkSpeedInfo in the HDD adapter
struct.

CRs-Fixed: 2207694
Bug: 111906763
Test: None
Change-Id: Ie0c84c271cee188e8bd1663095022daefd703f97