qcacld-2.0: Use request manager for tsm metrics
Propagation from qcacld-3.0 to qcacld-2.0

We are transitioning to the new request manager framework. Change
hdd_get_tsm_stats() and hdd_GetTsmStatsCB() to this framework. Note
that this framework provides the infrastructure to pass data from the
response thread to the request thread and hence eliminates the need to
maintain temporary tsmStats in the HDD adapter struct.

Change-Id: I0122bbdc7076fd5f4c80bb10e9340e68f1970bec
CRs-Fixed: 2207636
Test: None
Bug: 111906763