Don't build 8053 base SOM device tree with kernel
We are building apq8053-lite-som.dtb from source
in a custom rule so there is no need to build it
here. We also don't want it appended to the
kernel as it will be loaded from oem_bootloader
instead.

Also remove board-id and clean up formatting in
apq8053-lite-som.dts.

Bug: -
Test: apq8053-lite-som.dtb is not created.
Change-Id: I6a403196ad830e798cc335e683c244bc186f3ce8